Enable ReactPHP Support
